Moving entire timer code in timer.h
