+++ src/Makefile.am
@@ -107,7 +107,7 @@ liblol_core_sources = \
     mesh/mesh.cpp mesh/mesh.h \
     mesh/primitivemesh.cpp mesh/primitivemesh.h \
     \
-    sys/init.cpp sys/timer.cpp sys/file.cpp sys/hacks.cpp \
+    sys/init.cpp sys/file.cpp sys/hacks.cpp \
     sys/thread.cpp sys/threadtypes.cpp sys/getopt.cpp \
     \
     image/image.cpp image/image-private.h image/kernel.cpp image/pixel.cpp \
